+++ CHANGELOG.md
@@ -25,8 +25,8 @@ If you like this project and find it useful, please consider giving it a star on
 ## [2.1.1] - Dev branch
 
 - [package]: Apply uniform style.
-- [package]: Upgrade package.
 - [package]: Update dependencies.
+- [package]: Upgrade package: agents instructions and typescript 7.
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>